+++ SysManager/SysManager/ViewModels/DebloaterViewModel.cs
@@ -22,6 +22,9 @@ namespace SysManager.ViewModels;
 public sealed partial class DebloaterViewModel : ViewModelBase
 {
     /// <inheritdoc/>
+    protected internal override IRelayCommand? RefreshOnF5 => RefreshCommand;
+
+    /// <inheritdoc/>
     protected internal override IRelayCommand? EscapeCancel =>
         IsBusy ? CancelCommand : null;