+++ SysManager/SysManager/ViewModels/DuplicateFileViewModel.cs
@@ -27,6 +27,9 @@ namespace SysManager.ViewModels;
 public sealed partial class DuplicateFileViewModel : ViewModelBase
 {
     /// <inheritdoc/>
+    protected internal override IRelayCommand? RefreshOnF5 => ScanCommand;
+
+    /// <inheritdoc/>
     protected internal override IRelayCommand? EscapeCancel =>
         IsBusy ? CancelScanCommand : null;